+++ src/components/HappyTummy.jsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import {
   AppBar,
   Toolbar,
@@ -6,18 +6,30 @@ import {
   IconButton,
   Menu,
   MenuItem,
+  Box,
+  CircularProgress,
 } from "@mui/material";
 import AccountCircle from "@mui/icons-material/AccountCircle";
 import ManagerSection from "./ManagerSection";
 import LandingPage from "./LandingPage";
 import EmployeeSection from "./EmployeeSection";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
+import { getEmployees } from "../redux/thunks/allThunk";
 
 const HappyTummyApp = () => {
   const [role, setRole] = useState("");
   const [anchorEl, setAnchorEl] = useState(null);
 
-  const { employeesData } = useSelector((state) => state.all);
+  const dispatch = useDispatch();
+
+  const { employeesData, employeesDataStatus } = useSelector(
+    (state) => state.all
+  );
+
+  // API Call for get Employees
+  useEffect(() => {
+    dispatch(getEmployees());
+  }, [dispatch]);
 
   // Function to handle menu open/close
   const handleMenu = (event) => {
@@ -35,65 +47,83 @@ const HappyTummyApp = () => {
 
   return (
     <>
-      {/* Header */}
-      <AppBar position="sticky">
-        <Toolbar>
-          <Typography
-            variant="h6"
-            component="div"
-            sx={{ flexGrow: 1 }}
-            onClick={() => window.location.reload()}
-          >
-            {role ? `Happy Tummy (${role})` : `Happy Tummy`}
+      {employeesDataStatus === "loading" ? (
+        <Box
+          height={700}
+          display="flex"
+          flexDirection="column"
+          justifyContent="center"
+          alignItems="center"
+          gap={2}
+        >
+          <CircularProgress size={50} />
+          <Typography variant="h6">
+            It might take max. 2 mins to load the app
           </Typography>
-          <div>
-            <IconButton
-              size="large"
-              edge="end"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleMenu}
-              color="inherit"
-            >
-              <AccountCircle />
-            </IconButton>
-            <Menu
-              id="menu-appbar"
-              anchorEl={anchorEl}
-              anchorOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              open={Boolean(anchorEl)}
-              onClose={handleClose}
-            >
-              <MenuItem onClick={() => handleRoleSelect("Manager")}>
-                Manager
-              </MenuItem>
-              <MenuItem
-                onClick={() => handleRoleSelect("Employee")}
-                disabled={!employeesData?.length}
+        </Box>
+      ) : (
+        <>
+          {/* Header */}
+          <AppBar position="sticky">
+            <Toolbar>
+              <Typography
+                variant="h6"
+                component="div"
+                sx={{ flexGrow: 1 }}
+                onClick={() => window.location.reload()}
               >
-                Employee
-              </MenuItem>
-            </Menu>
-          </div>
-        </Toolbar>
-      </AppBar>
+                {role ? `Happy Tummy (${role})` : `Happy Tummy`}
+              </Typography>
+              <div>
+                <IconButton
+                  size="large"
+                  edge="end"
+                  aria-controls="menu-appbar"
+                  aria-haspopup="true"
+                  onClick={handleMenu}
+                  color="inherit"
+                >
+                  <AccountCircle />
+                </IconButton>
+                <Menu
+                  id="menu-appbar"
+                  anchorEl={anchorEl}
+                  anchorOrigin={{
+                    vertical: "top",
+                    horizontal: "right",
+                  }}
+                  keepMounted
+                  transformOrigin={{
+                    vertical: "top",
+                    horizontal: "right",
+                  }}
+                  open={Boolean(anchorEl)}
+                  onClose={handleClose}
+                >
+                  <MenuItem onClick={() => handleRoleSelect("Manager")}>
+                    Manager
+                  </MenuItem>
+                  <MenuItem
+                    onClick={() => handleRoleSelect("Employee")}
+                    disabled={!employeesData?.length}
+                  >
+                    Employee
+                  </MenuItem>
+                </Menu>
+              </div>
+            </Toolbar>
+          </AppBar>
 
-      {/* Show Landing Page when no option is Selected */}
-      {role === "" && <LandingPage />}
+          {/* Show Landing Page when no option is Selected */}
+          {role === "" && <LandingPage />}
 
-      {/* Show Manager Section if Manager is Selected */}
-      {role === "Manager" && <ManagerSection />}
+          {/* Show Manager Section if Manager is Selected */}
+          {role === "Manager" && <ManagerSection />}
 
-      {/* Show Employee Section if Employee is Selected */}
-      {role === "Employee" && <EmployeeSection />}
+          {/* Show Employee Section if Employee is Selected */}
+          {role === "Employee" && <EmployeeSection />}
+        </>
+      )}
     </>
   );
 };
